update products listing page to app router
