+++ app/api/product-categories/route.ts
@@ -1,18 +1,9 @@
-import { NextApiRequest } from "next";
 import { envIdCookieName, previewApiKeyCookieName } from "../../../lib/constants/cookies";
 import { getProductTaxonomy } from "../../../lib/kontentClient";
-import { parseBoolean } from "../../../lib/utils/parseBoolean";
-import { NextRequest } from "next/server";
-import { cookies } from "next/headers";
+import { cookies, draftMode } from "next/headers";
 
-export const GET = async (req: NextRequest) => {
-  const searchParams = req.nextUrl.searchParams;
-  const usePreview = parseBoolean(searchParams.get("preview"));
-
-  if (usePreview === null) {
-    return new Response("Please provide 'preview' query parameter with value 'true' or 'false'.", {status: 400});
-  }
-
+export const GET = async () => {
+  const usePreview = (await draftMode()).isEnabled;
   const cookiesList = await cookies();
   
   const currentEnvId = cookiesList.get(envIdCookieName);
